+++ src/components/chat/DyadWritePlan.tsx
@@ -1,13 +1,9 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { FileText, Eye, ChevronDown, ChevronUp, Loader2 } from "lucide-react";
-import { useSetAtom, useAtomValue } from "jotai";
-import { useQuery } from "@tanstack/react-query";
-import { previewModeAtom, selectedAppIdAtom } from "@/atoms/appAtoms";
-import { planStateAtom } from "@/atoms/planAtoms";
-import { selectedChatIdAtom } from "@/atoms/chatAtoms";
+import { useSetAtom } from "jotai";
+import { previewModeAtom } from "@/atoms/appAtoms";
 import { CustomTagState } from "./stateTypes";
-import { planClient } from "@/ipc/types/plan";
-import { queryKeys } from "@/lib/queryKeys";
+import { usePlan } from "@/hooks/usePlan";
 
 interface DyadWritePlanProps {
   node: {
@@ -25,46 +21,11 @@ export const DyadWritePlan: React.FC<DyadWritePlanProps> = ({ node }) => {
   const { title, summary, complete, state } = node.properties;
   const [showSummary, setShowSummary] = useState(false);
   const setPreviewMode = useSetAtom(previewModeAtom);
-  const chatId = useAtomValue(selectedChatIdAtom);
-  const appId = useAtomValue(selectedAppIdAtom);
-  const planState = useAtomValue(planStateAtom);
-  const setPlanState = useSetAtom(planStateAtom);
 
   // Consider in progress if state is pending OR complete is explicitly "false"
   const isInProgress = state === "pending" || complete === "false";
-  const hasPlanInMemory = chatId ? planState.plansByChatId.has(chatId) : false;
 
-  // Query for saved plan in .dyad/plans/ files
-  const { data: savedPlan } = useQuery({
-    queryKey: queryKeys.plans.forChat({
-      appId: appId ?? null,
-      chatId: chatId ?? null,
-    }),
-    queryFn: async () => {
-      if (!appId || !chatId) return null;
-      return planClient.getPlanForChat({ appId, chatId });
-    },
-    enabled: !!appId && !!chatId && !hasPlanInMemory && !isInProgress,
-    staleTime: 1000 * 60 * 5, // 5 minutes
-  });
-
-  // Load saved plan into memory state if found
-  useEffect(() => {
-    if (savedPlan && chatId && !hasPlanInMemory) {
-      setPlanState((prev) => {
-        const nextPlans = new Map(prev.plansByChatId);
-        nextPlans.set(chatId, {
-          content: savedPlan.content,
-          title: savedPlan.title,
-          summary: savedPlan.summary ?? undefined,
-        });
-        return {
-          ...prev,
-          plansByChatId: nextPlans,
-        };
-      });
-    }
-  }, [savedPlan, chatId, hasPlanInMemory, setPlanState]);
+  const { savedPlan, hasPlanInMemory } = usePlan({ enabled: !isInProgress });
 
   const hasPlan = hasPlanInMemory || !!savedPlan;
